requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
django-analytical
django-autocomplete-light
django-autoslug
django-bootstrap3
django-bower
django-bootstrap-breadcrumbs
django-cors-headers
django-countries
django-crispy-forms
django-debug-toolbar
django-environ
django-exclusivebooleanfield
django-filter
django_guardian
django-hats
django-hijack
django-hosts
django-mathfilters
django-rq
django-storages
django-tables2
django-taggit
django-taggit-serializer
django-test-addons
djangorestframework
djangorestframework-bulk